@@ -1,3 +1,3 @@
-import RBSheet from "./src";
+import RBSheet from './src';
 
 export default RBSheet;
