@@ -1,32 +1,32 @@
-import { StyleSheet } from "react-native";
+import {StyleSheet} from 'react-native';
 
 const styles = StyleSheet.create({
   wrapper: {
     flex: 1,
-    backgroundColor: "#00000077"
+    backgroundColor: '#00000077',
   },
   mask: {
     flex: 1,
-    backgroundColor: "transparent"
+    backgroundColor: 'transparent',
   },
   container: {
-    backgroundColor: "#fff",
-    width: "100%",
+    backgroundColor: '#fff',
+    width: '100%',
     height: 0,
-    overflow: "hidden"
+    overflow: 'hidden',
   },
   draggableContainer: {
-    width: "100%",
-    alignItems: "center",
-    backgroundColor: "transparent"
+    width: '100%',
+    alignItems: 'center',
+    backgroundColor: 'transparent',
   },
   draggableIcon: {
     width: 35,
     height: 5,
     borderRadius: 5,
     margin: 10,
-    backgroundColor: "#ccc"
-  }
+    backgroundColor: '#ccc',
+  },
 });
 
 export default styles;
